chore: refactor modules admin dashboard
