+++ modules/admin/articles/index.tsx
@@ -1,0 +1,12 @@
+import React from 'react'
+import Layout from '@/components/layouts/admin'
+
+const Article = () => {
+  return (
+    <Layout>
+      <h1>Articles</h1>
+    </Layout>
+  )
+}
+
+export default Article
