+++ modules/admin/dashboard/index.tsx
@@ -1,10 +1,11 @@
 import React from 'react'
+import Layout from '@/components/layouts/admin'
 
 const Dashboard = () => {
   return (
-    <div>
+    <Layout>
       <h1>Dashboard</h1>
-    </div>
+    </Layout>
   )
 }
