+++ modules/admin/projects/index.tsx
@@ -1,0 +1,12 @@
+import Layout from '@/components/layouts/admin'
+import React from 'react'
+
+const Projects = () => {
+  return (
+    <Layout>
+      <h1>Projects</h1>
+    </Layout>
+  )
+}
+
+export default Projects
